fix: rename culturalSurvey
