+++ src/features/culturalSurvey/api/__mocks__/useCulturalSurveyQuestionsQuery.ts
@@ -1,10 +1,10 @@
 import { UseQueryResult } from 'react-query'
 
 import { CulturalSurveyQuestionsResponse } from 'api/gen'
-import { useCulturalSurveyQuestions as actualUseCulturalSurveyQuestions } from 'features/culturalSurvey/api/useCulturalSurveyQuestions'
+import { useCulturalSurveyQuestionsQuery as actualUseCulturalSurveyQuestions } from 'features/culturalSurvey/api/useCulturalSurveyQuestionsQuery'
 import { culturalSurveyQuestionsFixture } from 'features/culturalSurvey/fixtures/culturalSurveyQuestions.fixture'
 
-export const useCulturalSurveyQuestions: typeof actualUseCulturalSurveyQuestions = jest.fn(
+export const useCulturalSurveyQuestionsQuery: typeof actualUseCulturalSurveyQuestions = jest.fn(
   () =>
     ({ data: culturalSurveyQuestionsFixture, isLoading: false, isSuccess: true }) as UseQueryResult<
       CulturalSurveyQuestionsResponse,
